+++ src/runUpdate.ts
@@ -1,57 +1,92 @@
 import chalk from "chalk";
+import equal from "deep-equal";
 
-import { getLocalPath, getWorkflowData, updateWorkflow } from "./manager";
+import { getWorkflowData, renderWorkflow, updateWorkflow } from "./manager";
+import { mergesToSteps, mergeWorkflows } from "./sanitizer";
 import { decapitalize } from "./utils";
 
 export default function runUpdate(
     name: string,
     localContent: string | null,
     remoteContent: string | null,
-    updateExisting: boolean
+    forceUpdate: boolean
 ): boolean {
-    const localPath = getLocalPath(name);
-    const workflowPrefix = `Workflow ${chalk.bold(localPath)}`;
     if (!remoteContent) {
-        console.warn(
-            chalk.red(`✗  ${workflowPrefix} download failed, skipping`)
-        );
+        console.warn(chalk.red("  ✗  download failed, skipping update"));
         return false;
     }
-    const workflowPurpose = chalk.bold(
-        decapitalize(getWorkflowData(remoteContent).name)
-    );
+    const remoteWorkflow = getWorkflowData(remoteContent);
+    const workflowPurpose = chalk.bold(decapitalize(remoteWorkflow.name));
     if (!localContent) {
         updateWorkflow(name, remoteContent);
-        console.info(
-            chalk.green(
-                `✓  Workflow ${chalk.bold(
-                    localPath
-                )} added to ${workflowPurpose}`
-            )
-        );
+        console.info(chalk.green(`  ✓  added to ${workflowPurpose}`));
         return true;
     }
-    if (localContent === remoteContent) {
+    const localWorkflow = getWorkflowData(localContent);
+    if (equal(localWorkflow, remoteWorkflow)) {
+        const cleanContent = renderWorkflow(localWorkflow);
+        if (cleanContent === localContent) {
+            console.log(
+                chalk.grey(`  ✓  up to date and ready to ${workflowPurpose}`)
+            );
+            return false;
+        }
+        updateWorkflow(name, cleanContent);
         console.log(
-            chalk.grey(
-                `✓  ${workflowPrefix} is up to date and ready to ${workflowPurpose}`
-            )
+            chalk.green(`  ✎  reformatted and ready to ${workflowPurpose}`)
         );
         return false;
     }
-    if (!updateExisting) {
-        console.info(
-            chalk.yellow(
-                `↻  ${workflowPrefix} can ${workflowPurpose} better, add ${chalk.bold(
-                    "-f"
-                )} CLI flag to update`
-            )
-        );
-        return false;
+
+    if (localWorkflow.name !== remoteWorkflow.name) {
+        if (forceUpdate) {
+            console.log(chalk.green("  ↻  Workflow name updated"));
+            localWorkflow.name = remoteWorkflow.name;
+        } else {
+            console.log(
+                "  ✎  Workflow name is different from remote, use --force flag to update"
+            );
+        }
     }
-    updateWorkflow(name, remoteContent);
+    if (!equal(localWorkflow.on, remoteWorkflow.on)) {
+        if (forceUpdate) {
+            console.log(chalk.green("  ↻  Workflow triggers updated"));
+            localWorkflow.on = remoteWorkflow.on;
+        } else {
+            console.log(
+                "  ✎  Workflow triggers are different from remote, use --force flag to update"
+            );
+        }
+    }
+    const stepMerges = mergeWorkflows(localWorkflow, remoteWorkflow);
+    stepMerges.forEach(stepMerge => {
+        if (stepMerge.action === "add") {
+            console.log(
+                chalk.green(`  ✓  Step ${chalk.bold(stepMerge.stepName)} added`)
+            );
+        }
+        if (stepMerge.action === "update") {
+            console.log(
+                chalk.green(
+                    `  ↻  Step ${chalk.bold(stepMerge.stepName)} updated`
+                )
+            );
+        }
+        if (stepMerge.action === "delete") {
+            console.log(
+                chalk.yellow(
+                    `  🗑  Step ${chalk.bold(stepMerge.stepName)} deleted`
+                )
+            );
+        }
+    });
+    localWorkflow.jobs = localWorkflow.jobs || remoteWorkflow.jobs || {};
+    const workflowJob = Object.values(localWorkflow.jobs || {})[0];
+    workflowJob.steps = mergesToSteps(stepMerges);
+    const renderedWorkflow = renderWorkflow(localWorkflow);
+    updateWorkflow(name, renderedWorkflow);
     console.info(
-        chalk.green(`✓  ${workflowPrefix} updated, cross-check changes`)
+        chalk.green(`  ✓  safely updated to ${chalk.bold(workflowPurpose)}`)
     );
     return true;
 }